+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.8
+
+### Added
+
+- Variabilization of timeout
+- Change backoff multiply 
+
+### Changed
+
 ## 0.1.0 (Unreleased)
 
 BACKWARDS INCOMPATIBILITIES / NOTES: